setting up controller
