+++ app/controllers/recipe_controller.rb
@@ -1,16 +1,14 @@
-require 'httparty'
 class RecipeController < ApplicationController
-  BASE_URL = "https://slack.com/api/"
-  TOKEN = ENV["SLACK_TOKEN"]
 
   def index
+
   end
 
   def show
-
+    @recpie = RecpieApiWrapper.recipe_details(params[:uri])
   end
 
   def result
-
+    @results = RecpieApiWrapper.recipe_query(params[:q])
   end
 end
